# 8286972: Support the new loop induction variable related PopulateIndex IR node on x86