--- a/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
+++ b/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
@@ -2276,6 +2276,7 @@ void C2_MacroAssembler::vectortest(int bt, int vlen, XMMRegister src1, XMMRegist
 
 void C2_MacroAssembler::vpadd(BasicType elem_bt, XMMRegister dst, XMMRegister src1, XMMRegister src2, int vlen_enc) {
   assert(UseAVX >= 2, "required");
+#ifdef ASSERT
   bool is_bw = ((elem_bt == T_BYTE) || (elem_bt == T_SHORT));
   bool is_bw_supported = VM_Version::supports_avx512bw();
   if (is_bw && !is_bw_supported) {
@@ -2283,6 +2284,7 @@ void C2_MacroAssembler::vpadd(BasicType elem_bt, XMMRegister dst, XMMRegister sr
     assert((dst->encoding() < 16) && (src1->encoding() < 16) && (src2->encoding() < 16),
            "XMM register should be 0-15");
   }
+#endif // ASSERT
   switch (elem_bt) {
     case T_BYTE: vpaddb(dst, src1, src2, vlen_enc); return;
     case T_SHORT: vpaddw(dst, src1, src2, vlen_enc); return;