This PR adds x86 backend support for the new loop induction variable related PopulateIndex IR node. 
This IR node was added as part of [JDK-8280510](https://bugs.openjdk.java.net/browse/JDK-8280510).

The performance numbers are as follows:
Before:
Benchmark                   (count)   Mode  Cnt       Score       Error  Units
IndexVector.exprWithIndex1    65536  thrpt    3   64556.552 ±  1126.396  ops/s
IndexVector.exprWithIndex2    65536  thrpt    3   22117.050 ± 11452.098  ops/s
IndexVector.indexArrayFill    65536  thrpt    3  117776.383 ±  1120.957  ops/s

After:
Benchmark                   (count)   Mode  Cnt       Score       Error  Units
IndexVector.exprWithIndex1    65536  thrpt    3  203180.290 ±  2147.807  ops/s
IndexVector.exprWithIndex2    65536  thrpt    3  274132.756 ±  6853.393  ops/s
IndexVector.indexArrayFill    65536  thrpt    3  374165.202 ± 46930.779  ops/s

Please review.

Best Regards,
Sandhya

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed (1 review required, with at least 1 reviewer)

### Issue
 * [JDK-8286972](https://bugs.openjdk.java.net/browse/JDK-8286972): Support the new loop induction variable related PopulateIndex IR node on x86


### Reviewers
 * [Vladimir Kozlov](https://openjdk.java.net/census#kvn) (@vnkozlov - **Reviewer**) ⚠️ Review applies to [727daec0](https://git.openjdk.java.net/jdk/pull/8778/files/727daec095e9779724fc5924dbf41d8c8de7c98f)
 * [Jatin Bhateja](https://openjdk.java.net/census#jbhateja) (@jatin-bhateja - Committer)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/8778/head:pull/8778` \
`$ git checkout pull/8778`

Update a local copy of the PR: \
`$ git checkout pull/8778` \
`$ git pull https://git.openjdk.java.net/jdk pull/8778/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 8778`

View PR using the GUI difftool: \
`$ git pr show -t 8778`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/8778.diff">https://git.openjdk.java.net/jdk/pull/8778.diff</a>

</details>